Ported more utility functions
